+++ src/main/kotlin/wisp/perseanchronicles/telos/pt3_arrow/MenriSystemCreator.kt
@@ -11,7 +11,7 @@ import wisp.questgiver.wispLib.placeInSector
 import java.awt.Color
 
 object MenriSystemCreator {
-    val systemName = "Lama"
+    val systemBaseName = "Lama"
 
     fun createMenriSystem(): PlanetAPI? {
         // gate
@@ -21,12 +21,12 @@ object MenriSystemCreator {
         val systemOrbitDays = 180f
         val planetName = "Menri"
 
-        if (game.sector.getStarSystem(systemName) != null) {
+        if (game.sector.getStarSystem(systemBaseName) != null) {
             game.logger.w { "Lama system already exists!" }
-            return game.sector.getStarSystem(systemName).planets.firstOrNull { it.name == planetName }
+            return game.sector.getStarSystem(systemBaseName).planets.firstOrNull { it.name == planetName }
         }
 
-        val system = game.sector.createStarSystem(systemName)
+        val system = game.sector.createStarSystem(systemBaseName)
 
         system.backgroundTextureFilename = "graphics/backgrounds/background5.jpg"
         system.addTag(Tags.THEME_HIDDEN)
@@ -40,7 +40,7 @@ object MenriSystemCreator {
         )
             .apply {
                 spec
-                name = systemName
+                name = systemBaseName
                 applySpecChanges()
             }